refactor: simplify App, MainArea, TopPanel, and MenuSection components by removing unused state and props for cleaner code
