+++ src/App.tsx
@@ -1,4 +1,3 @@
-import { useState, useEffect, useRef } from 'react'
 import { PictRunner } from '@takeyaqa/pict-browser'
 import { Analytics } from './components'
 import {
@@ -6,7 +5,6 @@ import {
   NotificationMessageSection,
   FooterSection,
 } from './sections'
-import { Result } from './types'
 import { ConfigProvider } from './features/config'
 import { ModelProvider } from './features/model'
 import { MainArea } from './layouts'
@@ -16,27 +14,6 @@ interface AppProps {
 }
 
 function App({ pictRunnerInjection }: AppProps) {
-  const [result, setResult] = useState<Result | null>(null)
-  const [pictRunnerLoaded, setPictRunnerLoaded] = useState(false)
-  const pictRunner = useRef<PictRunner>(null)
-
-  useEffect(() => {
-    // Use the injected PictRunner for testing
-    if (pictRunnerInjection) {
-      pictRunner.current = pictRunnerInjection
-      setPictRunnerLoaded(true)
-      return
-    }
-    const loadPictRunner = async () => {
-      pictRunner.current = new PictRunner()
-      await pictRunner.current.init()
-      setPictRunnerLoaded(true)
-    }
-
-    // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    loadPictRunner()
-  }, [pictRunnerInjection])
-
   return (
     <>
       <HeaderSection />
@@ -45,12 +22,7 @@ function App({ pictRunnerInjection }: AppProps) {
       />
       <ConfigProvider>
         <ModelProvider>
-          <MainArea
-            pictRunnerLoaded={pictRunnerLoaded}
-            pictRunner={pictRunner}
-            result={result}
-            setResult={setResult}
-          />
+          <MainArea pictRunnerInjection={pictRunnerInjection} />
         </ModelProvider>
       </ConfigProvider>
       <FooterSection />
